requirements.txt
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
breadability==0.1.20
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
docopt==0.6.2
docx==0.2.4
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
Flask-Session==0.8.0
frontend==0.0.3
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
msgspec==0.19.0
multidict==6.4.4
mysql-connector==2.2.9
mysql-connector-python==9.2.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
openai==1.84.0
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
prov==2.0.1
psycopg2==2.9.10
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.0
PyJWT==2.10.1
PyMuPDF==1.26.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.0
sumy==0.11.0
tk==0.1.0
tqdm==4.67.1
traits==7.0.2
twilio==9.6.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
Werkzeug==3.1.3
yarl==1.20.0
